backend/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
flask_sqlalchemy
pymysql
mysql-connector-python
sqlalchemy-utils
flask_cors
pytest
coverage
Flask-Session
redis
sphinx
pytest-timeout
flake8